@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-//<snippet29>
+﻿//<snippet29>
 namespace CancelParallelLoops
 {
     using System;
@@ -15,30 +10,32 @@ namespace CancelParallelLoops
     {
         static void Main()
         {
-            int[] nums = Enumerable.Range(0, 10000000).ToArray();
-            CancellationTokenSource cts = new CancellationTokenSource();
+            int[] nums = Enumerable.Range(0, 10_000_000).ToArray();
+            CancellationTokenSource cts = new();
 
-           // Use ParallelOptions instance to store the CancellationToken
-            ParallelOptions po = new ParallelOptions();
-            po.CancellationToken = cts.Token;
-            po.MaxDegreeOfParallelism = System.Environment.ProcessorCount;
+            // Use ParallelOptions instance to store the CancellationToken
+            ParallelOptions options = new()
+            {
+                CancellationToken = cts.Token,
+                MaxDegreeOfParallelism = Environment.ProcessorCount
+            };
             Console.WriteLine("Press any key to start. Press 'c' to cancel.");
             Console.ReadKey();
 
             // Run a task so that we can cancel from another thread.
             Task.Factory.StartNew(() =>
             {
-                if (Console.ReadKey().KeyChar == 'c')
+                if (Console.ReadKey().KeyChar is 'c')
                     cts.Cancel();
                 Console.WriteLine("press any key to exit");
             });
 
             try
             {
-                Parallel.ForEach(nums, po, (num) =>
+                Parallel.ForEach(nums, options, (num) =>
                 {
                     double d = Math.Sqrt(num);
-                    Console.WriteLine("{0} on {1}", d, Thread.CurrentThread.ManagedThreadId);
+                    Console.WriteLine("{0} on {1}", d, Environment.CurrentManagedThreadId);
                 });
             }
             catch (OperationCanceledException e)
