@@ -1,42 +1,39 @@
 ﻿namespace ThreadLocalFor
 {
-//<snippet05>
- using System;
- using System.Linq;
- using System.Threading;
- using System.Threading.Tasks;
+    //<snippet05>
+    using System;
+    using System.Linq;
+    using System.Threading;
+    using System.Threading.Tasks;
 
- class Test
- {
-     static void Main()
-     {
-         int[] nums = Enumerable.Range(0, 1_000_000).ToArray();
-         long total = 0;
+    class Test
+    {
+        static void Main()
+        {
+            int[] nums = Enumerable.Range(0, 1_000_000).ToArray();
+            long total = 0;
 
-         // Use type parameter to make subtotal a long, not an int
-         Parallel.For<long>(0, nums.Length, () => 0, (j, loop, subtotal) =>
-         {
-             subtotal += nums[j];
-             return subtotal;
-         },
-             subtotal => Interlocked.Add(ref total, subtotal)
-         );
+            // Use type parameter to make subtotal a long, not an int
+            Parallel.For<long>(0, nums.Length, () => 0,
+                (j, loop, subtotal) =>
+                {
+                    subtotal += nums[j];
+                    return subtotal;
+                },
+                subtotal => Interlocked.Add(ref total, subtotal));
 
-         Console.WriteLine("The total is {0:N0}", total);
-         Console.WriteLine("Press any key to exit");
-         Console.ReadKey();
-     }
- }
-//</snippet05>
+            Console.WriteLine("The total is {0:N0}", total);
+            Console.WriteLine("Press any key to exit");
+            Console.ReadKey();
+        }
+    }
+    //</snippet05>
 }
 
 namespace NotInUse
 {
     using System;
-    using System.Drawing;
-    using System.IO;
     using System.Linq;
-    using System.Threading;
     using System.Threading.Tasks;
 
     class FromSnippet1
@@ -56,7 +53,7 @@ namespace NotInUse
 
         static void ComputeHeron(double x, int precision)
         {
-            Random rand = new Random();
+            Random rand = Random.Shared;
 
             double r = rand.Next(1, (int)x);
             double f = (double)1 / precision;
